first work version

+++ app/src/main/java/com/valab/exaltcalculation/PriceObject.kt
@@ -1,5 +1,5 @@
 package com.valab.exaltcalculation
 
 object PriceObject {
-    var exaltPrice:Float = 0f
+    var exaltPrice:Double = 0.0
 }
